Criado a interface PacienteService e a classe PacienteServiceImpl
